♻️ refactor:  ドメイン層のフォルダ構成変更
Issue #145
