+++ lib/domain/entities/owner.dart
