+++ lib/infrastructure/github/api.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a MIT license that can be
 // found in the LICENSE file.
 
-import '../../domain/repo/repositories/repo_repository.dart';
+import '../../domain/repositories/repo_repository.dart';
 
 /// GitHub API の定義
 /// 各APIはUriを返す
