+++ lib/infrastructure/github/repo_repository.dart
@@ -4,7 +4,7 @@
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../domain/repo/repositories/repo_repository.dart';
+import '../../domain/repositories/repo_repository.dart';
 import 'api.dart';
 import 'exception.dart';
 import 'http_client.dart';
