+++ lib/infrastructure/hive/app_data_repository.dart
@@ -5,8 +5,8 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 
-import '../../domain/app_data/repositories/app_data_repository.dart';
-import '../../domain/repo/repositories/repo_repository.dart';
+import '../../domain/repositories/app_data_repository.dart';
+import '../../domain/repositories/repo_repository.dart';
 
 /// Hive版アプリデータRepositoryプロバイダー
 final hiveAppDataRepositoryProvider = Provider<HiveAppDataRepository>(
