+++ lib/presentation/repo/components/selected_repo.dart
@@ -6,8 +6,8 @@ import 'package:equatable/equatable.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 
-import '../../../domain/repo/entities/repo.dart';
-import '../../../domain/repo/repositories/repo_repository.dart';
+import '../../../domain/entities/repo.dart';
+import '../../../domain/repositories/repo_repository.dart';
 import '../../../utils/logger.dart';
 import '../view_page.dart';
