+++ test/domain/app_data/repositories/app_data_repository_test.dart
@@ -6,7 +6,7 @@ import 'dart:io';
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:github_search/domain/app_data/repositories/app_data_repository.dart';
+import 'package:github_search/domain/repositories/app_data_repository.dart';
 
 import '../../../test_utils/hive.dart';
 import '../../../test_utils/locale.dart';
