+++ test/domain/repo/entities/repo_test.dart
@@ -3,7 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:github_search/domain/repo/entities/repo.dart';
+import 'package:github_search/domain/entities/repo.dart';
 import 'package:github_search/infrastructure/github/json_object/repo/repo.dart';
 import 'package:hexcolor/hexcolor.dart';
