+++ test/infrastructure/hive/app_data_repository_test.dart
@@ -5,7 +5,7 @@
 import 'dart:io';
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:github_search/domain/repo/repositories/repo_repository.dart';
+import 'package:github_search/domain/repositories/repo_repository.dart';
 import 'package:github_search/infrastructure/hive/app_data_repository.dart';
 
 import '../../test_utils/hive.dart';
