+++ test/presentation/repo/components/selected_repo_test.dart
@@ -6,8 +6,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:github_search/config/router.dart';
-import 'package:github_search/domain/repo/entities/owner.dart';
-import 'package:github_search/domain/repo/entities/repo.dart';
+import 'package:github_search/domain/entities/owner.dart';
+import 'package:github_search/domain/entities/repo.dart';
 import 'package:github_search/presentation/repo/components/selected_repo.dart';
 import 'package:github_search/presentation/repo/view_page.dart';
 import 'package:go_router/go_router.dart';
